requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.6
django-ckeditor==6.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.12.4
environ==1.0
filelock==3.3.0
gunicorn==19.9.0
idna==3.2
oauthlib==3.1.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2018.5
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.8.1
